Update grammar (int/state distinction), rearrange files
