+++ examples/nutshells/roed.ruel
@@ -169,7 +169,7 @@ FF6000: 37..53  # d, yellow
 
 
 @ICONS
-?  000 FFF [38]
+?  000 .. FFF [38]
 0: 303030
 1: FF0011
 2: 00AAFF
