--- nutshell/segment_types/table/lark_assets/__init__.py
@@ -1,3 +1,0 @@
-from .transformer import *
-from ._exceptions import *
-from ._classes import *
